serial-receiver-py/pyproject.toml
[project]
name = "serial-receiver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyserial>=3.5",
]
